+++ bootique-logback-sentry/src/test/java/io/bootique/logback/sentry/LogbackSentryContextFactoryIT.java
@@ -33,6 +33,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+@Deprecated
 public class LogbackSentryContextFactoryIT {
 
     @Test