+++ bootique-logback-sentry/src/test/java/io/bootique/logback/sentry/LogbackSentryModuleTest.java
@@ -23,6 +23,7 @@ import io.bootique.junit5.BQModuleTester;
 import io.bootique.junit5.BQTest;
 import org.junit.jupiter.api.Test;
 
+@Deprecated
 @BQTest
 public class LogbackSentryModuleTest {